+++ transport/stringifyError.ts
